move swap_workspace logic to swap command
