+++ sway/sway.5.scd
@@ -344,18 +344,20 @@ set|plus|minus|toggle <amount>
 	"Sticks" a floating window to the current output so that it shows up on all
 	workspaces.
 
-*swap* container with id|con_id|mark <arg>
-	Swaps the position, geometry, and fullscreen status of two containers. The
-	first container can be selected either by criteria or focus. The second
-	container can be selected by _id_, _con_id_, or _mark_. _id_ can only be
-	used with xwayland views. If the first container has focus, it will retain
-	focus unless it is moved to a different workspace or the second container
-	becomes fullscreen on the same workspace as the first container. In either
-	of those cases, the second container will gain focus.
+*swap* (container with id|con_id|mark <arg>) |
+	(workspace with [number] <number>)
 
-*swap_workspace* with [number] <name>
-	Swaps the content of the currently focused workspace with the content
-	of the workspace specified as an argument.
+	If using the container keyword, swaps the position, geometry, and
+	fullscreen status of two containers. The first container can be selected
+	either by criteria or focus. The second container can be selected by
+	_id_, _con_id_, or _mark_. _id_ can only be used with xwayland views.
+	If the first container has focus, it will retain focus unless it is moved
+	to a different workspace or the second container becomes fullscreen on the
+	same workspace as the first container. In either of those cases, the
+	second container will gain focus.
+
+	If using the workspace keyword, swaps the content of the currently focused
+	workspace with the content of the workspace specified as an argument.
 
 *title_format* <format>
 	Sets the format of window titles. The following placeholders may be used:
